requirements.txt
python-dotenv
openai
requests
beautifulsoup4
pyYAML